Unnecessary comments deleted. Xpath shortened. Login and new task tests put in one file. Main file renamed.
