+++ test/e2e/sourceryTests.js
@@ -9,42 +9,39 @@ module.exports = {
 
         let currentUser = "Demo User";
         let currentUserRole = "Admin";
-        let newTaskName = "RimantasTest2";
-        let newTaskDescription = "newTaskCreationTestDescription-2";
-        let newTaskBillToClient = "Yes";
-        let newTaskHourlyRate = "10";
+        let menuItems = ['Time Logging','Invoices','Tasks','Projects','Clients','Time Entries'];
         browser
         .url(browser.launchUrl)
-        .waitForElementVisible(common.pageTitle); // wait for the Login title
+        .waitForElementVisible(common.pageTitle); 
         //Click to expand select user dropdown
         browser.element('css selector',login.userSelect , function(result) {
             if(result.status === c.ELEMENT_FOUND) { 
                 browser.click(login.userSelect);
             }
         });
-        //Select from expanded droprown
+
         browser.element('css selector', login.getSpecificSelectUserOption(currentUser), function(result) {
             if(result.status === c.ELEMENT_FOUND) { 
                 browser.click('css selector', login.getSpecificSelectUserOption(currentUser));
             }
         });
+
         browser.assert.containsText(login.userSelectedValue, currentUser);
 
-        //Click to expand select role dropdown
         browser.element('css selector', login.userRoleSelectField, function(result) {
             if(result.status === c.ELEMENT_FOUND) { 
                 browser.click('css selector', login.userRoleSelectField);
             }
         });
-        //Select from expanded droprown
+
         browser.element('css selector', login.getSpecificSelectUserRoleOption(currentUserRole), function(result) {
             if(result.status === c.ELEMENT_FOUND) { 
                 browser.click('css selector', login.getSpecificSelectUserRoleOption(currentUserRole));
             }
         });
+
         browser.assert.containsText(login.userRoleSelectField, currentUserRole);
 
-        //Click submit button
         browser.element('css selector', common.submitButton, function(result) {
             if(result.status === c.ELEMENT_FOUND) {
                 browser
@@ -52,9 +49,29 @@ module.exports = {
                 .waitForElementVisible(common.loggedInUsersName);
             }
         });
-        browser.assert.containsText(common.loggedInUsersName, currentUser);
 
-        //Click menu list element "tasks"
+         browser.assert.containsText(common.loggedInUsersName, currentUser)
+                .saveScreenshot(conf.imgpath(browser) + 'Demo.png');
+
+        //assert main nav elements are displayed
+        browser.assert.containsText(common.menuList, menuItems[0])
+                .assert.containsText(common.menuList, menuItems[1])
+                .assert.containsText(common.menuList, menuItems[2])
+                .assert.containsText(common.menuList, menuItems[3])
+                .assert.containsText(common.menuList, menuItems[4])
+                .assert.containsText(common.menuList, menuItems[5])
+        //assert menu item "time logging" is selected and marked in blue
+                .assert.containsText(common.menuListActiveElement, menuItems[0])
+                .assert.cssProperty(common.menuListActiveElement,'color','rgba(64, 76, 237, 1)'); //color blue         
+    },
+
+    'create new task': function (browser) {
+
+        let newTaskName = "RimantasTest2"+Date.now();
+        let newTaskDescription = "newTaskCreationTestDescription-2";
+        let newTaskBillToClient = "Yes";
+        let newTaskHourlyRate = "10";
+       
         browser.element('css selector',common.menuListTasksElement , function(result) {
             if(result.status === c.ELEMENT_FOUND) { 
                  browser
@@ -62,8 +79,7 @@ module.exports = {
                  .waitForElementVisible(tasks.createTaskButton);
             }
         });
-        
-        //Click "create task" button
+               
         browser.element('css selector',tasks.createTaskButton , function(result) {
             if(result.status === c.ELEMENT_FOUND) { 
                  browser
@@ -71,8 +87,7 @@ module.exports = {
                  .waitForElementVisible(tasks.createTaskFormTitle);
             }
         });    
-        
-        //select "Task name" input field and set value
+                
         browser.element('css selector', tasks.createTaskName, function(result) {
             if(result.status === c.ELEMENT_FOUND) { 
                 browser
@@ -81,7 +96,6 @@ module.exports = {
             }
         });
 
-        //select "Task description" input field and set value
         browser.element('css selector', tasks.createTaskDescription, function(result) {
             if(result.status === c.ELEMENT_FOUND) { 
                 browser
@@ -90,7 +104,7 @@ module.exports = {
             }
         });
 
-        //select "bill to client" dropdown
+        //select "bill to client" input field
         browser.element('css selector', tasks.createTaskBillToClient, function(result) {
             if(result.status === c.ELEMENT_FOUND) { 
                 browser
@@ -105,7 +119,6 @@ module.exports = {
             }
         });
 
-         //select "Hourly rate (USD)" input field and set value
          browser.element('css selector', tasks.createTaskHourlyRate, function(result) {
             if(result.status === c.ELEMENT_FOUND) { 
                 browser
@@ -114,14 +127,12 @@ module.exports = {
             }
         });
 
-        //Click "Save" button
         browser.element('css selector',tasks.createTaskSaveButton , function(result) {
             if(result.status === c.ELEMENT_FOUND) { 
                  browser.click(tasks.createTaskSaveButton);
             }
         });
 
-        //Click menu list element "tasks"
         browser.element('css selector',common.menuListTasksElement , function(result) {
             if(result.status === c.ELEMENT_FOUND) { 
                  browser
@@ -130,7 +141,6 @@ module.exports = {
             }
         });
 
-        //click on task description filter
         browser.element('xpath', tasks.tasksNameFilter , function(result) {
             if(result.status === c.ELEMENT_FOUND) { 
                  browser
